refactor(auto-update)!: skip files that could be ignored
The `auto-update` flag will no longer cause new paths to be added to the
registry of ignored files. This prevents regressions from sneaking in,
when some file was already conforming to the stricter config and an
ignored error code started appearing in that file.

Fixes https://github.com/Gelio/loose-ts-check/issues/16